# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Projektor API",
    "description": "Self-hosted media server for movies, TV shows, and anime.",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "LibraryKindInput": {
        "type": "string",
        "enum": ["movie", "tv", "anime"]
      },
...
      "CreateLibraryRequestInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "kind": {
            "$ref": "#/components/schemas/LibraryKindInput"
          },
          "paths": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "required": ["name", "kind", "paths"]
      },
...
      "LibraryKind": {
        "type": "string",
        "enum": ["movie", "tv", "anime"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["statusCode", "error", "message"],
        "additionalProperties": false
      },
...
      "Library": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "description": "Opaque identifier"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "kind": {
            "$ref": "#/components/schemas/LibraryKind"
          },
          "paths": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "description": "ISO 8601 UTC timestamp"
          },
          "lastScannedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                "description": "ISO 8601 UTC timestamp"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["id", "name", "kind", "paths", "createdAt", "lastScannedAt"],
        "additionalProperties": false
      },
...
      "ScanStatus": {
        "type": "object",
        "properties": {
          "libraryId": {
            "type": "string",
            "minLength": 1,
            "description": "Opaque identifier"
          },
          "state": {
            "type": "string",
            "enum": ["idle", "running"]
          },
          "filesSeen": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "filesChanged": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "filesMissing": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "filesProbed": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "filesFailed": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Files ffprobe could not read"
          },
          "itemsLinked": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Files linked to a movie or episode this scan"
          },
          "itemsMatched": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Items matched to TMDB this scan"
          },
          "itemsUnmatched": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Items TMDB search could not match confidently"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                "description": "ISO 8601 UTC timestamp"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "libraryId",
          "state",
          "filesSeen",
          "filesChanged",
          "filesMissing",
          "filesProbed",
          "filesFailed",
          "itemsLinked",
          "itemsMatched",
          "itemsUnmatched",
          "startedAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/libraries": {
      "get": {
        "summary": "List libraries",
        "tags": ["libraries"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Library"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a library (admin)",
        "tags": ["libraries"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibraryRequestInput"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{id}": {
      "get": {
        "summary": "Get a library",
        "tags": ["libraries"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Opaque identifier"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a library and everything indexed from it (admin)",
        "tags": ["libraries"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Opaque identifier"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{id}/scan": {
      "post": {
        "summary": "Scan a library now (admin). Runs to completion before responding.",
        "tags": ["libraries"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Opaque identifier"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanStatus"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}